feat(preprocessor): adjust package main entry to replaces
When using e.g. typescript files you also reference the .ts file in the
main entry of the package.json. The given preprocessor to pkg could
transform the file and will replace the extension regarding the given
pkg config. The file reference in the main entry has to be adjusted too.